+++ graphbrain/agents/claim_actors.py
@@ -1,5 +1,6 @@
 from graphbrain.meaning.actors import find_actors
 from graphbrain.agents.agent import Agent
+from graphbrain.agents.system import wrap_edge
 
 
 class ClaimActors(Agent):
@@ -17,4 +18,5 @@ class ClaimActors(Agent):
         _, main_actor, claim, main_edge = edge
         actors = find_actors(self.hg, claim)
         for actor in actors:
-            self.add(('claim-actor/P/.', main_actor, actor, claim, main_edge))
+            yield wrap_edge(
+                ('claim-actor/P/.', main_actor, actor, claim, main_edge))