+++ graphbrain/agents/claims.py
@@ -1,10 +1,11 @@
 from collections import Counter
 import progressbar
-from graphbrain import *
-from graphbrain.meaning.concepts import *
+from graphbrain import hedge
+from graphbrain.meaning.concepts import strip_concept, has_proper_concept
 from graphbrain.meaning.lemmas import deep_lemma
 from graphbrain.meaning.corefs import main_coref
 from graphbrain.agents.agent import Agent
+from graphbrain.agents.system import wrap_edge
 
 
 CLAIM_PRED_LEMMAS = {'say', 'claim'}
@@ -138,7 +139,7 @@ class Claims(Agent):
                 # write gender
                 if gender:
                     gender_atom = '{}/P/.'.format(gender)
-                    self.add((gender_atom, actor))
+                    yield wrap_edge((gender_atom, actor))
 
                 i += 1
                 bar.update(i)
@@ -174,7 +175,7 @@ class Claims(Agent):
                         self.anaphoras += 1
 
                 # write claim
-                self.add(('claim/P/.', actor, claim, edge))
+                yield wrap_edge(('claim/P/.', actor, claim, edge))
 
                 i += 1
                 bar.update(i)