+++ graphbrain/agents/conflicts.py
@@ -1,9 +1,12 @@
-from graphbrain import *
-from graphbrain.meaning.concepts import *
+from graphbrain import hedge
+from graphbrain.meaning.concepts import (all_concepts,
+                                         strip_concept,
+                                         has_proper_concept)
 from graphbrain.meaning.lemmas import deep_lemma
 from graphbrain.meaning.corefs import main_coref
 from graphbrain.meaning.actors import is_actor
 from graphbrain.agents.agent import Agent
+from graphbrain.agents.system import wrap_edge
 
 
 CONFLICT_PRED_LEMMAS = {'warn', 'kill', 'accuse', 'condemn', 'slam', 'arrest',
@@ -34,8 +37,8 @@ class Conflicts(Agent):
                 if item[0].to_str() in CONFLICT_TOPIC_TRIGGERS:
                     for concept in all_concepts(item[1]):
                         if self.hg.degree(concept) > 1:
-                            self.add(('conflict-topic/P/.', actor_orig,
-                                      actor_targ, concept, edge))
+                            yield wrap_edge(('conflict-topic/P/.', actor_orig,
+                                             actor_targ, concept, edge))
                             self.conflict_topics += 1
 
     def input_edge(self, edge):
@@ -60,8 +63,10 @@ class Conflicts(Agent):
                                 ('conflict/P/.', actor_orig, actor_targ, edge))
                             if (is_actor(self.hg, actor_orig) and
                                     is_actor(self.hg, actor_targ)):
-                                self.add(conflict_edge)
-                                self._topics(actor_orig, actor_targ, edge)
+                                yield wrap_edge(conflict_edge)
+                                for wedge in self._topics(
+                                        actor_orig, actor_targ, edge):
+                                    yield wedge
                                 self.conflicts += 1
 
     def report(self):