+++ graphbrain/agents/taxonomy.py
@@ -1,5 +1,6 @@
 import graphbrain.constants as const
 from graphbrain.agents.agent import Agent
+from graphbrain.agents.system import wrap_edge
 
 
 class Taxonomy(Agent):
@@ -26,4 +27,4 @@ class Taxonomy(Agent):
                     parent = edge[1]
                 if parent:
                     ont_edge = (const.type_of_pred, edge, parent)
-                    self.add(ont_edge, primary=False)
+                    yield wrap_edge(ont_edge, primary=False)